==> 02-data.Rmd <==
# Data sources

## EIA Data
This [dataset](https://github.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/blob/main/data/raw/monthly_fuel_generation_by_state_full.csv) was pulled from the U.S. Energy Information Administration [EIA website](https://www.eia.gov/opendata/qb.php?category=1) through an API. There is a package in R called "eia" that helps automate data pulls from the EIA.gov API calls. We used this package to pull monthly power generation data by fuel type for all of the states in the USA from the time period of 2001 - 2021.
```{r, include = F, message = F, warning = F}
library(tidyverse)
library(kableExtra)
eia_dat <- read.csv('data/raw/monthly_fuel_generation_by_state_full.csv')
```

### Column Descriptions
The columns in this dataset are listed below: 

* **thousand_megawatthours:** net power generation in thousand megawatt hours
* **date:** contains the month in which the power was generated. This column ranges from "2001-01-01" to "2021-08-01".
* **state:** contains the state in which the power was generated. Unique values include all 50 USA states
* **fuel:** contains the fuel type, with unique values listed below:

```{r}
unique(eia_dat$fuel)
```
The fuels are aggregated hierarchically as outlined below:

* **All Fuels:** includes the aggregation of power generation for coal, petroleum liquids, petroleum coke, natural gas, other gases, nuclear, conventional hydroelectric, hydro-electric pumped storage, other, and other renewables (total).
* **Other renewables (total):** includes the aggregation of power generation for wind, all utility-scale solar, geothermal, and biomass (total).
* **Biomass (total):** includes the aggregation of power generation for wood and wodd-derive fuels and other biomass.
* **All solar:** includes the aggregation of power generation for small-scale solar photovoltaic and all utility-scale solar. 

The column summaries of data types and statistics (for numeric columns) are shown below:
```{r}
summary(eia_dat)
```
The dataset has dimensions of:
```{r}
dim(eia_dat)
```
or 145239 rows and 4 columns. 

This dataset only included power generation data for fuel types that were actually produced in a state. Therefore, some fuel type generation values are missing from certain states. Additionally, some states only adopted certain fuel types at a specific time interval during the period of interest, so again the power generation data is missing for some of the time period. This will be discussed more in the [missing values chapter](https://edav-ccfall21-final-project-team-7.github.io/USCleanEnergyAnalysis/missing-values.html)


## WB Data
The World Bank [country dataset](https://github.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/blob/main/data/clean/clean_worldbank.csv) for the U.S. is downloaded from [Data Bank - World Development Indicators](https://databank.worldbank.org/country/USA/556d8fa6/Popular_countries#) in CSV File format. The dataset captured annual statistics for 48 macro socio-economic indicators (rows) from 1960 to 2020 (columns). The below table describes the short names we used for these development indicators in our data analysis for the ease of data visualization.   

```{r, message = F, warning = F}
library(tidyverse)
library(kableExtra)
wb_dat <- read.csv('data/clean/WorldBankDataDictionary.csv')
names(wb_dat) <- c('Column Name', 'Description')
wb_dat %>% kable(escape = F) %>% 
  kable_styling("striped", font_size = 9, full_width = F) %>% 
  scroll_box(width = "500px", height = "300px")

```

## Financial Data
This dataset was obtained from Yahoo Finance website. The prices (closing, opening etc) for different energy related Exchange Traded Funds (ETFs) which are similar to S&P500 or NASDAQ like [XLE](https://raw.githubusercontent.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/main/data/raw/XLE.csv), (overall energy sector fund), [CNRG](https://raw.githubusercontent.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/main/data/raw/CNRG.csv) (clean power ETF), and [ICLN](https://raw.githubusercontent.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/main/data/raw/ICLN.csv) (global clean energy ETF) were downloaded as a CSV file and added to the dataset folder. The date range for every ETF varies from the day the ETF went public to the day the data was downloaded (10th November 2021). However, the daily prices of crude oil were obtained from the West Texas Intermediate ([WTI](https://raw.githubusercontent.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/main/data/clean/clean_crushing_crude_oil.csv)) which was obtained from the following [website](https://www.macrotrends.net/1369/crude-oil-price-history-chart).


The table below shows the XLE ETF which is the energy sector fund. All other ETFs datasets have the exact same columns with different dates and prices. 
```{r, message = F, warning = F}
xle <- read_csv('data/raw/xle.csv')
xle %>% kable(escape = F) %>% 
  kable_styling("striped", font_size = 9, full_width = F) %>% 
  scroll_box(width = "500px", height = "300px")
```


The table below shows the crude oil dataset which contains the daily price of crude oil up to 8th November 2021
```{r, message = F, warning = F}
oil <- read_csv('data/clean/clean_crushing_crude_oil.csv')
oil %>% kable(escape = F) %>% 
  kable_styling("striped", font_size = 9, full_width = F) %>% 
  scroll_box(width = "500px", height = "300px")
```

### Column Descriptions

The columns in the XLE (and all ETF) dataset are listed below: 

* **date:** contains the date of every trading day (non-holidays and non-weekends) from the time the ETF went public to the date data was downloaded ("2021-11-10")
* **open:** contains the opening price of the ETF for that day
* **high:** contains the highest price of the ETF for that day
* **low:** contains the lowest price of the ETF for that day
* **close:** contains the close price of the ETF for that day
* **adj close:** contains the adjusted close price of the ETF for that day after adjustment from payments and dividends
* **volume:** contains the volume of ETF traded that day


The columns in the crude oil dataset are listed below: 

* **day:** contains the date in the range from "01/02/1986" to "11/08/2021"
* **cost per barrel:** contains the cost per barrel of oil based on West Texas Intermediate (WTI)
