==> 03-cleaning.Rmd <==
# Data transformation

## EIA Data
### Corrections
There were several issues with the EIA data. The clean version of the data can be found [here](https://github.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/blob/main/data/clean/clean_monthly_fuel_generation.csv). First, the state name read in as "USA-<state>" (examples below).

```{r, warning=F, echo=F}
library(tidyverse)
library(kableExtra)
library(lubridate) 
df <- read.csv('data/raw/monthly_fuel_generation_by_state_full.csv')
unique(df$state)[1:5]

```

We updated this column by splitting on the "-" delimiter and only keeping the actual state name. 

```{r, warning=F, echo=F}
df <- df %>% separate(state, sep = "-", into = c('country','state'), remove = TRUE)
df$country <- NULL
unique(df$state)[1:5]
```

Other formatting issues existed with the fuel type. There was white space at the start and end of each fuel and some fuels had a space in the middle of it (i.e. natural gas ) which is shown below. 

```{r, warning=F, echo=F}
unique(df$fuel)[1:5]
```

To make these variables easier to work with, we trimmed the white space at the beginning and end of the fuels and added an underscore if there was a space in the middle of the fuel type. A sample of the updated fuels is shown below:

```{r, warning=F, echo=F}
df$fuel <- trimws(df$fuel) 
df$fuel <- gsub(" ", "_", df$fuel)
df$fuel <- gsub("-", "_", df$fuel)
unique(df$fuel)[1:5]
```

The "date" column of the data was imported as a character column, so we updated it to be a date datatype.

```{r, warning=F, echo=F}
df$date <- as.Date(df$date)
```

Lastly, there was an issue where the solar generation total was not aggregated correctly on the EIA website. This is detailed more in the [missing values chapter](https://edav-ccfall21-final-project-team-7.github.io/USCleanEnergyAnalysis/missing-values.html). To fix this, we performed a pivot on the data and recalculated the aggregation, ignoring N/a's in the data. We also imputed zero for other values of the power generation so that aggregations were not affected in later code. 

```{r, warning=F, echo=F}
df[is.na(df$thousand_megawatthours),]$thousand_megawatthours <- 0
#Update solar summation to be correct
df <- df %>% pivot_wider(id_cols = c(date, state), names_from=c(fuel), values_from = thousand_megawatthours)
df <- df %>% rowwise() %>% mutate(all_solar = sum(small_scale_solar_photovoltaic, all_utility_scale_solar, na.rm = TRUE))
df <- df %>% pivot_longer(cols=names(df)[3:21],  names_to='fuel', values_to='thousand_megawatthours')
df[is.na(df$thousand_megawatthours),]$thousand_megawatthours <- 0

# i<-0
# for (vals in row.names(df)) {
#   i<-i+1
#   df$state[i] <- str_split(df$state[i],"-",simplify=TRUE)[2]
# }

# write.csv(df,'data/clean/clean_monthly_fuel_generation.csv',row.names = FALSE)
```

### Create interactive data

The data needed to be transformed for the interactive component of this project. The data is wrangled to create [dataframes](https://github.com/EDAV-CCFall21-Final-Project-Team-7/USCleanEnergyAnalysis/tree/main/data/clean) (listed as d2001.csv - d2021.csv) that contain the renewable percent of total generation for each state for each year as shown below. As mentioned in the data introduction, the total power generation is the aggregate of coal, petroleum_liquids, petroleum_coke, natural_gas, other_gases, nuclear, conventional_hydroelectric, other, and other_renewables_(total) fuel types. Additionally, the renewables value is the sum of the conventional_hydroelectric and other_renewables_(total) fuels. After performing these aggregations, the percent of renewables was found by dividing the aggregate renewables by the total power generation for each state on each date. 

```{r, warning=F, echo=F}
df <- read.csv('data/clean/clean_monthly_fuel_generation.csv')
# interactive_data <- df %>% group_by(state, fuel, date) %>% pivot_wider(id_cols = c(state, date), names_from= fuel, values_from = thousand_megawatthours)
# interactive_data[is.na(interactive_data)] <- 0
# interactive_data <- interactive_data[names(interactive_data) %in% c('state', 'date', 'coal', 'petroleum_liquids','petroleum_coke','natural_gas', 'other_gases', 'nuclear', 'conventional_hydroelectric', 'other', 'other_renewables_(total)')]
# 
# interactive_data_percent <- interactive_data %>%
#   gather(variable, value, -c(state,date)) %>%
#   group_by(state,date) %>%
#   mutate(percentage = value/sum(value, na.rm=TRUE)*100) %>%
#   select(-value) %>%
#   spread(variable, percentage)
# interactive_data_percent$renewables <- interactive_data_percent$conventional_hydroelectric + interactive_data_percent$`other_renewables_(total)`
# write.csv(interactive_data_percent, 'data/clean/interactive_data.csv',row.names = FALSE)


df$year <- year(df$date)
df$date <- NULL
interactive_data<-df %>%
  group_by(fuel,year,state) %>%
  summarize(thousand_megawatthours = sum(thousand_megawatthours,na.rm=TRUE)) %>% 
  pivot_wider(id_cols = c(state, year), names_from= fuel, values_from = thousand_megawatthours)

interactive_data <- interactive_data[names(interactive_data) %in% c('state', 'year', 'coal', 'petroleum_liquids','petroleum_coke','natural_gas', 'other_gases', 'nuclear', 'conventional_hydroelectric', 'other', 'other_renewables_(total)')]

interactive_data_percent <- interactive_data %>%
  gather(variable, value, -c(state,year)) %>%
  group_by(state,year) %>%
  mutate(percentage = value/sum(value, na.rm=TRUE)*100) %>%
  select(-value) %>%
  spread(variable, percentage)
interactive_data_percent$renewables <- interactive_data_percent$conventional_hydroelectric + interactive_data_percent$`other_renewables_(total)`

interactive_data_percent <- interactive_data_percent[names(interactive_data_percent) %in% c('state','year','renewables')]

head(interactive_data_percent, 5)
# write.csv(interactive_data_percent, 'data/clean/interactive_data.csv',row.names = FALSE)
```

```{r}
d2001<-interactive_data_percent %>% filter(year=='2001')
d2001$year<-NULL

d2002<-interactive_data_percent %>% filter(year=='2002')
d2002$year<-NULL

d2003<-interactive_data_percent %>% filter(year=='2003')
d2003$year<-NULL

d2004<-interactive_data_percent %>% filter(year=='2004')
d2004$year<-NULL

d2005<-interactive_data_percent %>% filter(year=='2005')
d2005$year<-NULL

d2006<-interactive_data_percent %>% filter(year=='2006')
d2006$year<-NULL

d2007<-interactive_data_percent %>% filter(year=='2007')
d2007$year<-NULL

d2008<-interactive_data_percent %>% filter(year=='2008')
d2008$year<-NULL

d2009<-interactive_data_percent %>% filter(year=='2009')
d2009$year<-NULL

d2010<-interactive_data_percent %>% filter(year=='2010')
d2010$year<-NULL

d2011<-interactive_data_percent %>% filter(year=='2011')
d2011$year<-NULL

d2012<-interactive_data_percent %>% filter(year=='2012')
d2012$year<-NULL

d2013<-interactive_data_percent %>% filter(year=='2013')
d2013$year<-NULL

d2014<-interactive_data_percent %>% filter(year=='2014')
d2014$year<-NULL

d2015<-interactive_data_percent %>% filter(year=='2015')
d2015$year<-NULL

d2016<-interactive_data_percent %>% filter(year=='2016')
d2016$year<-NULL

d2017<-interactive_data_percent %>% filter(year=='2017')
d2017$year<-NULL

d2018<-interactive_data_percent %>% filter(year=='2018')
d2018$year<-NULL

d2019<-interactive_data_percent %>% filter(year=='2019')
d2019$year<-NULL

d2020<-interactive_data_percent %>% filter(year=='2020')
d2020$year<-NULL

d2021<-interactive_data_percent %>% filter(year=='2021')
d2021$year<-NULL

names(d2001)<-c("month","value")
names(d2002)<-c("month","value")
names(d2003)<-c("month","value")
names(d2004)<-c("month","value")
names(d2005)<-c("month","value")
names(d2006)<-c("month","value")
names(d2007)<-c("month","value")
names(d2008)<-c("month","value")
names(d2009)<-c("month","value")
names(d2010)<-c("month","value")
names(d2011)<-c("month","value")
names(d2012)<-c("month","value")
names(d2013)<-c("month","value")
names(d2014)<-c("month","value")
names(d2015)<-c("month","value")
names(d2016)<-c("month","value")
names(d2017)<-c("month","value")
names(d2018)<-c("month","value")
names(d2019)<-c("month","value")
names(d2020)<-c("month","value")
names(d2021)<-c("month","value")

d2001<-d2001[order(-d2001$value,d2001$month),]
d2002<-d2002[order(-d2002$value,d2002$month),]
d2003<-d2003[order(-d2003$value,d2003$month),]
d2004<-d2004[order(-d2004$value,d2004$month),]
d2005<-d2005[order(-d2005$value,d2005$month),]
d2006<-d2006[order(-d2006$value,d2006$month),]
d2007<-d2007[order(-d2007$value,d2007$month),]
d2008<-d2008[order(-d2008$value,d2008$month),]
d2009<-d2009[order(-d2009$value,d2009$month),]
d2010<-d2010[order(-d2010$value,d2010$month),]
d2011<-d2011[order(-d2011$value,d2011$month),]
d2012<-d2012[order(-d2012$value,d2012$month),]
d2013<-d2013[order(-d2013$value,d2013$month),]
d2014<-d2014[order(-d2014$value,d2014$month),]
d2015<-d2015[order(-d2015$value,d2015$month),]
d2016<-d2016[order(-d2016$value,d2016$month),]
d2017<-d2017[order(-d2017$value,d2017$month),]
d2018<-d2018[order(-d2018$value,d2018$month),]
d2019<-d2019[order(-d2019$value,d2019$month),]
d2020<-d2020[order(-d2020$value,d2020$month),]
d2021<-d2021[order(-d2021$value,d2021$month),]


# write.csv(d2001, 'draft code/d2001',row.names = FALSE)
# write.csv(d2002, 'draft code/d2002',row.names = FALSE)
# write.csv(d2003, 'draft code/d2003',row.names = FALSE)
# write.csv(d2004, 'draft code/d2004',row.names = FALSE)
# write.csv(d2005, 'draft code/d2005',row.names = FALSE)
# write.csv(d2006, 'draft code/d2006',row.names = FALSE)
# write.csv(d2007, 'draft code/d2007',row.names = FALSE)
# write.csv(d2008, 'draft code/d2008',row.names = FALSE)
# write.csv(d2009, 'draft code/d2009',row.names = FALSE)
# write.csv(d2010, 'draft code/d2010',row.names = FALSE)
# write.csv(d2011, 'draft code/d2011',row.names = FALSE)
# write.csv(d2012, 'draft code/d2012',row.names = FALSE)
# write.csv(d2013, 'draft code/d2013',row.names = FALSE)
# write.csv(d2014, 'draft code/d2014',row.names = FALSE)
# write.csv(d2015, 'draft code/d2015',row.names = FALSE)
# write.csv(d2016, 'draft code/d2016',row.names = FALSE)
# write.csv(d2017, 'draft code/d2017',row.names = FALSE)
# write.csv(d2018, 'draft code/d2018',row.names = FALSE)
# write.csv(d2019, 'draft code/d2019',row.names = FALSE)
# write.csv(d2020, 'draft code/d2020',row.names = FALSE)
# write.csv(d2021, 'draft code/d2021',row.names = FALSE)
```

## WB Data
### Cleaning Data 
The worldbank dataset contains extra columns that are of not any use for the purpose of this project such as the 'country name', 'country code', and 'series code'. This is because we are looking at only one country and the 'Series Name' represents the same information series code would represent. It should also be noted that the column names for the years are redundantly long and it repeats the same things. The missing cells in the dataframe are filling with a string and they need to be replaced by NA so that they appear as missing data. The table below shows the raw data before any processing.
```{r} 
worldbank <-read_csv('data/raw/US_Country_Data_WorldBank.csv') 
worldbank %>% kable(escape = F) %>% 
  kable_styling("striped", font_size = 9, full_width = F) %>% 
  scroll_box(width = "800px", height = "300px")
```
After the above mentioned things have been corrected (changing column names and adding NAs to the data) and reduendent columns have been dropped the dataset looks like the following,


```{r, warning=F, echo=F}

i<-0
for (val in names(worldbank)){
  i<-i+1
  names(worldbank)[i] <- str_split(names(worldbank)[i]," ",simplify=TRUE)[1]
}
worldbank <- worldbank[,2:length(names(worldbank))]
names(worldbank)[3] <- 'Series Code'

worldbank <- worldbank %>% filter(Country == 'USA')

worldbank<-as.data.frame(worldbank)


for (name in names(worldbank)){
  for (row in row.names(worldbank)){
    if (worldbank[as.integer(row),name] == ".."){
      worldbank[as.integer(row),name] <- NA
    }
  }
}

drop <- c("Series Code","Country")

worldbank<-worldbank[,!(names(worldbank) %in% drop)]

# write.csv(worldbank,'data/clean/clean_worldbank.csv',row.names = FALSE)

worldbank %>% kable(escape = F) %>% 
  kable_styling("striped", font_size = 9, full_width = F) %>% 
  scroll_box(width = "800px", height = "300px")
```


## Financial Data
### Correcting Date format
In the dataset for oil data there were different types of date formats such as 'MM/DD/YY' and 'MM/DD/YYYY' as shown below.
```{r}
library(tidyverse)
library(kableExtra)
library(lubridate) 

Crushing_crude_oil <-read_csv('data/raw/Cushing_OK_WTI_Spot_Price_FOB.csv')
head(Crushing_crude_oil,10)
```

We updated the dataset to contain dates in only one format ("MM/DD/YYYY"). This was done by splitting the date based on "/" as the delimiter. 
```{r}
str_split(Crushing_crude_oil$Day[1],"/",simplify=TRUE)
```
We then checked if the date was already in the correct format by checking the number of characters in the year component of the split. If it was not correct then we added the missing components to the year format and joined the string again while updating that in the dataframe. If the year component in the was an integer less than 22 then we added "20" before the date making the year "20xx" but if the integer was greater than 22, it is obvious that it cannot be a date from the 21th century therefore we added "19" before the date making it "19xx".

```{r}

i<-0
for(val in Crushing_crude_oil$Day){
  i<-i+1
  if(nchar(str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[3]) == 2){
    temp = paste(str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[1],
                 str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[2],
                 sep = "/")
    if(as.integer(str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[3])<22)     {
      temp2 = paste("20",
                  str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[3],
                  sep="")
    }
    else{
      temp2 = paste("19",
                  str_split(Crushing_crude_oil$Day[i],"/",simplify=TRUE)[3],
                  sep="")
    }
    
    Crushing_crude_oil$Day[i]<-paste(temp,temp2,sep="/")
  }
}
# write.csv(Crushing_crude_oil,'data/clean/clean_crushing_crude_oil.csv',
#           row.names=FALSE)
head(Crushing_crude_oil,10)
```
