==> 05-results.Rmd <==
# Results
## Renewable Fuel Adoption
The below chart explores the average percent of renewable power generated in 2020 by each state. Renewable energy sources include wind, solar, geothermal, biomass, and hydroelectric power sources. We can see that Iowa is using mainly wind as its renewable energy source; whereas Hawaii is using mainly solar power. 

```{r, warning=F, echo=F}
library(tidyverse)
library(GGally)
library(lubridate)
library(forecast)
library(gridExtra)
library(cowplot)
library(RColorBrewer)
library(usmap)
library(patchwork)
library(ggpmisc)
df <- read.csv('data/clean/clean_monthly_fuel_generation.csv')
# df$fuel <- trimws(df$fuel) 
df$date <- as.Date(df$date)
# df[is.na(df$thousand_megawatthours),]$thousand_megawatthours <- 0
# df$fuel <- gsub(" ", "_", df$fuel)
# df$fuel <- gsub("-", "_", df$fuel)
#Update solar summation to be correct
# df <- df %>% pivot_wider(id_cols = c(date, state), names_from=c(fuel), values_from = thousand_megawatthours)
# df <- df %>% rowwise() %>% mutate(all_solar = sum(small_scale_solar_photovoltaic, all_utility_scale_solar, na.rm = TRUE))
# df <- df %>% pivot_longer(cols=names(df)[3:21],  names_to='fuel', values_to='thousand_megawatthours')
```

```{r, warning=F, echo=F, fig.height=8}
current <- df %>% filter(year(date) == 2020)
current_states <- current %>% group_by(state, fuel) %>% summarize(total_thousand_mwh = mean(thousand_megawatthours, na.rm=TRUE)) %>% pivot_wider(id_cols = c(state), names_from= fuel, values_from = total_thousand_mwh)
current_states[is.na(current_states)] <- 0
# current_states <- current_states[names(current_states) %in% c('state', 'coal', 'wind','all_solar', 'petroleum_liquids','petroleum_coke','natural_gas', 'other_gases', 'nuclear', 'conventional_hydroelectric', 'other', 'other_renewables_(total)')]
current_states$total <- current_states$coal + current_states$petroleum_coke + current_states$petroleum_liquids + current_states$natural_gas + current_states$other_gases + current_states$nuclear + current_states$conventional_hydroelectric + current_states$other + current_states$wind + current_states$all_solar + current_states$geothermal + current_states$other_biomass + current_states$wood_and_wood_derived_fuels 
current_states$renewables <- current_states$conventional_hydroelectric + current_states$wind + current_states$all_solar + current_states$geothermal + current_states$other_biomass + current_states$wood_and_wood_derived_fuels 

current_states$coal <- current_states$coal/current_states$total
current_states$renewables <- current_states$renewables/current_states$total
current_states$all_solar <- current_states$all_solar/current_states$total
current_states$wind <- current_states$wind/current_states$total


# current_states_percent <- current_states %>% 
#   gather(variable, value, -state) %>% 
#   group_by(state) %>% 
#   mutate(percentage = value/sum(value, na.rm=TRUE)*100) %>% 
#   select(-value) %>% 
#   spread(variable, percentage)

plot_df <- current_states[names(current_states) %in% c('state', 'all_solar', 'wind', 'renewables')]
plot_df <- plot_df %>% pivot_longer(cols = !state, names_to = "fuel", values_to = "percent") %>% mutate(fuel = replace(fuel, fuel == "all_solar", "solar"))

plot_df$fuel <- fct_relevel(plot_df$fuel, c('renewables', 'wind', 'solar'))


ggplot(plot_df, aes(x = percent, y = fct_reorder2(state, fuel == "renewables", percent, .desc = FALSE),
       color = fuel)) +
  geom_point() +
  scale_x_continuous(labels = scales::percent) +
  ggtitle("2020 Average Renewable Percent of Total Power Generation") +
  xlab("Net Generation (%)") +
  ylab("State") + 
  theme_linedraw(base_size = 12) +
  theme(plot.title = element_text(hjust=.5))
options(repr.P.width=4,repr.P.height=10)

```

It is interesting to note that Vermont is already generating 100% of their power from renewable energy sources. Additionally, Delaware has the lower percentage of power generated from renewables with these sources comprising just over 2% of the average 2020 power.

Some states have set [Renewable Portfolio Goals (RPS)](https://www.ncsl.org/research/energy/renewable-portfolio-standards.aspx) that specify that utilities must sell a certain percentage of electricity that is generated from renewable sources. In their RPS, Delaware set a goal in 2005 to achieve achieve power generation that is 25% renewables by 2025; however, they have only achieved 2% renewables in the current year. 

## Macro Trends
To explore macro trends in the energy sources utilized by the United States as a whole, the time series for each power source is plotted from 2001 - 2021. It is seen below that in 2001 coal was the predominant electricity source for America. Coal had a steady power generation until it began to decline around 2010. Natural gas took off at this time and by 2021 is the biggest producer of energy. Conventional hydroelectric was one of the first renewable energy resources to appear at significant levels in the electricity power generation for the US. Both solar and wind power have slowly increased in the amount of power they produce. 

```{r, echo=F, warning=F, fig.width = 10}
df  %>% 
  group_by(fuel, date) %>% 
  summarise(mil_mwhrs = sum(thousand_megawatthours, na.rm = TRUE)/1000) %>% 
  filter(fuel %in% c('all_solar', 'coal','natural_gas', 'wind', 'nuclear', 'conventional_hydroelectric')) %>%
  mutate(fuel = replace(fuel, fuel == "conventional_hydroelectric", "hydroelectric")) %>% 
  mutate(fuel = replace(fuel, fuel == "all_solar", "solar")) %>% 
  ggplot(aes(x = date, y = mil_mwhrs, color = fct_relevel(fuel, c('natural_gas', 'coal','nuclear', 'hydroelectric', 'wind', 'solar')))) +
  geom_line() +
  ggtitle('USA Net Generation by Fuel Type') + 
  xlab("Date") +
  ylab("MW Hours (Mil)") +
  theme_classic(12)+
  theme(legend.title=element_blank(), plot.title = element_text(hjust=.5))
```

```{r, echo=F, warning=F}
# df  %>% filter(fuel %in% c('wind','coal', 'natural_gas', 'all_solar')) %>%
#   select(c(fuel, date, thousand_megawatthours)) %>% 
#   group_by(fuel, date) %>% 
#   summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE)) %>% ungroup()  %>%
#   pivot_wider(names_from=fuel, id_cols = date, values_from = thousand_megawatthours) %>%
#   ggparcoord(columns=c(array(2:5)), scale= "uniminmax", alphaLines = .3, splineFactor = 9) +
#   theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1),
#         plot.title = element_text(hjust=.5)) +
#   ggtitle('Relationship between Fuel Types') +
#   xlab("Fuel") +
#   ylab("") 

```

We can explore the relationship between specific fuel types to determine which are correlated. From the plot below, we can see that solar is negatively correlated with coal and positively correlated with natural gas and wind. This is not necessarily surprising because natural gas is typically used to complement variable renewable energy sources like solar and wind and act as a form of stability to produce power when the sun isn't shining or the wind isn't blowing. It makes sense that as the adoption of wind and solar energy increases, natural gas usage will also increase until a cleaner alternative is found. Natural gas is a cleaner fuel source than coal and produces 50% less CO2 emissions according to the [Met Group](https://group.met.com/fyouture/natural-gas-vs-coal/66). However, the cleanest alternative would be another renewable energy source that could help address the variability in production of wind and solar, like batteries.  

```{r, echo=F, warning=F}
df  %>% filter(fuel %in% c('wind','coal', 'natural_gas', 'all_solar', 'nuclear')) %>%
  group_by(fuel, date) %>% summarise(million_mwhrs = sum(thousand_megawatthours, na.rm = TRUE)/1000) %>%
  pivot_wider(names_from=fuel, id_cols = c(date), values_from = million_mwhrs) %>% ungroup() %>% select(-c( date)) %>% ggpairs(lower=list(continuous=wrap(ggally_smooth, color='darkgray', size=1)),
  diag=list(continuous=wrap(ggally_barDiag, color='black')), 
  upper=list(axisLabels='show',continuous=wrap('cor', color='black'))) + 
  xlab('MW Hours (Mil)') +
  ylab('MW Hours (Mil)') +
  theme_bw() +
  theme(legend.position = "none", 
        panel.grid.major = element_blank(), 
        axis.ticks = element_blank(), 
        panel.border = element_rect( colour = "black", fill = NA),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1),
        plot.title = element_text(hjust=.5)) +
  ggtitle('Correlation of Fuel Source Power Generation')
```

To explore the seasonal component of different fuels we looked at the seasonplot for 2015-2021, which was when solar was adopted more widely as a fuel type. The solar power net generation seasonal trend peaks in the summer months (May - August) and dips in the winter months. This is most likely due to the fact that there is more sunshine in the summer months, so more solar power is able to be generated. The coal plot shows the power generation trends overall. There is a peak in June - September and December - February. These peaks make sense because these are the times of year that energy demand is at the highest due to heating demand during the cold months and cooling demand during the hot months. In the seasonal plot, we can see that solar is increasing in production from 2015-2021, while coal is decreasing in production levels.

```{r, warning=F, echo =F}
my_colors <- RColorBrewer::brewer.pal(9, "Greens")[3:9]
seasonal_plot_data <- df  %>% filter(fuel %in% c('wind','coal', 'natural_gas', 'all_solar', 'nuclear')) %>%
  group_by(fuel, date) %>% summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE)) %>%
  pivot_wider(names_from=fuel, id_cols = c(date), values_from = thousand_megawatthours) %>% ungroup()
seasonal_plot_data <- seasonal_plot_data %>% mutate(month_val = month(date))
seasonal_plot_data$date <- as.Date(seasonal_plot_data$date)

seasonal_plot_data <- seasonal_plot_data %>% filter(year(seasonal_plot_data$date) > 2014)
solar_series <- ts(seasonal_plot_data$all_solar/1000, start = c(2015,1,1), frequency = 12)
wind_series <- ts(seasonal_plot_data$wind/1000, start = c(2015,1,1), frequency = 12)
coal_series <- ts(seasonal_plot_data$coal/1000, start = c(2015,1,1), frequency = 12)
g1 <- ggseasonplot(solar_series) +
  labs(title = "Solar") +
  xlab("Month") + 
  ylab("MW Hours (Mil)") +
  guides(colour = guide_legend(reverse = TRUE)) +
  theme_bw(12) +
  theme(plot.title = element_text(hjust = 0.5)) + 
  scale_color_manual(values=my_colors) +
  theme(legend.position = "none")

g2 <- ggseasonplot(coal_series) +
  labs(title = "Coal") +
  xlab("Month") + 
  ylab("MW Hours (Mil)") +
  guides(colour = guide_legend(reverse = TRUE)) +
  theme_bw(12) +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_color_manual(values=my_colors) +
  theme(legend.position = "none")

gplaceholder <- ggseasonplot(coal_series) +
  guides(colour = guide_legend(reverse = TRUE)) +
  theme_bw(12) +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_color_manual(values=my_colors) +
  theme(legend.position = "right")

extract_legend <- function(my_ggp) {
  step1 <- ggplot_gtable(ggplot_build(my_ggp))
  step2 <- which(sapply(step1$grobs, function(x) x$name) == "guide-box")
  step3 <- step1$grobs[[step2]]
  return(step3)
}

# Apply user-defined function to extract legend
shared_legend <- extract_legend(gplaceholder)

#grid.arrange(arrangeGrob(g1,shared_legend,g2, nrow=2, ncol=2) ,widths = c(10, 1), top=textGrob("Seasonal #Plot for Total Power Generation", gp=gpar(fontsize=16,font=8)))

plot_row <- plot_grid(g1,shared_legend, g2, align = "v", nrow = 2, rel_widths  = c(7/8, 1/8, 7/8))

title <- ggdraw() + 
  draw_label(
    "Seasonal Plot for Total Power Generation by Fuel",
    x = .5,
    hjust = .5
  ) +
  theme(
    # add margin on the left of the drawing canvas,
    # so title is aligned with left edge of first plot
    plot.margin = margin(0, 0, 0, 7)
  )
plot_grid(
  title, plot_row,
  ncol = 1,
  # rel_heights values control vertical title margins
  rel_heights = c(0.1, 1)
)
#https://wilkelab.org/cowplot/articles/plot_grid.html - cowplot
#shared legend function: https://statisticsglobe.com/add-common-legend-to-combined-ggplot2-plots-in-r/

```

Next, we explore the primary fuel source for the USA states in 2001 and 2020. The most noticeable change in primary fuels can be seen from coal and natural gas. In 2001, 32 states used coal as the primary fuel, while by 2020 this number changed to 11. Natural gas experienced the opposite trend where it grew from being the primary fuel for 7 states to 22. Nuclear, hydroelectric, and petroleum liquids were also primary fuels for states in 2001 and 2020. These fuels did not see much change in number of states using them as their primary fuel. Renewables (other than hydroelectric) did become the primary fuel for 4 states (Iowa, Kansas, Maine, and Minnesota) in 2020, while in 2001 no states were using other renewables as the primary fuel.

```{r, echo=F, warning=F}
max_fuel_2020 <- df %>% filter(year(date) == 2020) %>%
  select(c(fuel, state, thousand_megawatthours)) %>% 
  group_by(fuel, state) %>% 
  summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE)) %>% ungroup()  %>%
  group_by(state) %>% filter(thousand_megawatthours == max(thousand_megawatthours)) 

max_fuel_2020 <- max_fuel_2020 %>% group_by(fuel) %>% summarize(count=n())
max_fuel_2020$year <- '2020'

max_fuel_2001 <- df %>% filter(year(date) == 2001) %>%
  select(c(fuel, state, thousand_megawatthours)) %>% 
  group_by(fuel, state) %>% 
  summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE)) %>% ungroup()  %>%
  group_by(state) %>% filter(thousand_megawatthours == max(thousand_megawatthours)) 

max_fuel_2001 <- max_fuel_2001 %>% group_by(fuel) %>% summarize(count=n())
max_fuel_2001$year <- '2001'

full_df <- rbind(max_fuel_2020, max_fuel_2001)
full_df$fuel[full_df$fuel == "conventional_hydroelectric"] <- 'hydroelectric'
full_df$fuel[full_df$fuel == "other_renewables_(total)"] <- 'other_renewables'
fuel <- 'other_renewables'
count <- 0
year <- '2001'
missing_row <- data.frame(fuel, count,year)
full_df <- rbind(full_df, missing_row)

ggplot(full_df, aes(x = fct_reorder2(fuel, year == '2001', count), y = count, fill =year)) +
  geom_col(position = "dodge") +
  xlab("Primary Fuel") +
  ylab("State Count") +
  ggtitle("Count of States' Primary Fuel Source") +
  theme_classic() +
  theme(plot.title = element_text(hjust=.5), axis.text.x = element_text(angle = 45, hjust=1)) + scale_fill_brewer(palette = "Paired")

```

The heatmaps below provide insights into how the energy production is distributed over the United States. It can be noticed that California is the major producer of solar energy whereas Texas is the major producer of coal, natural nas, and wind energy. 

```{r,fig.width=16,fig.height=16}
df_all_solar <- df  %>% 
  group_by(fuel, date, state) %>% summarise(thousand_megawatthours = sum(thousand_megawatthours)) %>% filter(fuel %in% c('all_solar'))

df_coal <- df  %>% 
  group_by(fuel, date, state) %>% summarise(thousand_megawatthours = sum(thousand_megawatthours)) %>% filter(fuel %in% c('coal'))

df_natural_gas <- df  %>% 
  group_by(fuel, date, state) %>% summarise(thousand_megawatthours = sum(thousand_megawatthours)) %>% filter(fuel %in% c('natural_gas'))

df_wind <- df  %>% 
  group_by(fuel, date, state) %>% summarise(thousand_megawatthours = sum(thousand_megawatthours)) %>% filter(fuel %in% c('wind'))

df_all_solar1 <- df_all_solar %>% group_by(state) %>% summarize(mil_mwhrs = mean(thousand_megawatthours)/1000) 

df_coal1 <- df_coal %>% group_by(state) %>% summarize(mil_mwhrs = mean(thousand_megawatthours)/1000) 

df_natural_gas1 <- df_natural_gas %>% group_by(state) %>% summarize(mil_mwhrs = mean(thousand_megawatthours)/1000) 

df_wind1 <- df_wind %>% group_by(state) %>% summarize(mil_mwhrs = mean(thousand_megawatthours)/1000) 



g1 <- plot_usmap(data = df_all_solar1, values = 'mil_mwhrs',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Avg. MW Hrs \n(Mil)", label = scales::comma) + 
  labs(title = "Solar") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g2 <- plot_usmap(data = df_coal1, values = 'mil_mwhrs',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Avg. MW Hrs \n(Mil)", label = scales::comma) + 
  labs(title = "Coal") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g3 <- plot_usmap(data = df_natural_gas1, values = 'mil_mwhrs',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Avg. MW Hrs \n(Mil)", label = scales::comma) + 
  labs(title = "Natural Gas") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g4 <- plot_usmap(data = df_wind1, values = 'mil_mwhrs',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Avg. MW Hrs \n(Mil)", label = scales::comma) + 
  labs(title = "Wind") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())



g1+g2+g3+g4+plot_layout(ncol=2,nrow=2,widths=c(8,8),heights=c(38,38))+plot_annotation(title = "Raw Values for Monthly Average megawatthours (Mil)",theme = theme(plot.title = element_text(size = 36,face='bold', hjust=.5)))
```

However, we know that the size of every state is the United States varies a lot. Therefore, to account for the size and provide a better comparison of the energy produced in a state, we normalized the amount of energy produced by the area of the state and plotted it on a heatmap as done above. It can be noticed from this heatmap that Texas is no longer one of the dominating producer of coal and natural gas. In fact, Texas has one of the highest values for wind energy per square mile, following closely behind Iowa and Oklahoma. New Jersey, Massachusetts, and California have the most solar energy produced per square mile. West Virginia, Indiana, and Ohio are producing the most coal energy per square mile. 

```{r,fig.width=16,fig.height=16}
area <- as.data.frame(read_csv('data/raw/area.csv'))

df_all_solar2 <- merge(df_all_solar1,area,by='state')
df_all_solar2$thousand_megawatthours <- df_all_solar2$mil_mwhrs*1000000/df_all_solar2$area

df_coal2 <- merge(df_coal1,area,by='state')
df_coal2$thousand_megawatthours<-df_coal2$mil_mwhrs*1000000/df_coal2$area

df_natural_gas2 <- merge(df_natural_gas1,area,by='state')
df_natural_gas2$thousand_megawatthours<-df_natural_gas2$mil_mwhrs*1000000/df_natural_gas2$area

df_wind2 <- merge(df_wind1,area,by='state')
df_wind2$thousand_megawatthours<-df_wind2$mil_mwhrs*1000000/df_wind2$area


g1 <- plot_usmap(data = df_all_solar2, values = 'thousand_megawatthours',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Normalized Avg \nMW Hrs", label = scales::comma) + 
  labs(title = "Solar") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g2 <- plot_usmap(data = df_coal2, values = 'thousand_megawatthours',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name ="Normalized Avg \nMW Hrs", label = scales::comma) + 
  labs(title = "Coal") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g3 <- plot_usmap(data = df_natural_gas2, values = 'thousand_megawatthours',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Normalized Avg \nMW Hrs", label = scales::comma) + 
  labs(title = "Natural Gas") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())

g4 <- plot_usmap(data = df_wind2, values = 'thousand_megawatthours',regions = 'states', labels = TRUE, color = 'pink')+
  scale_fill_continuous(low = "pink", high = "red", name = "Normalized Avg \nMW Hrs", label = scales::comma) + 
  labs(title = "Wind") +
  theme_classic(16) +
  xlab('') +
  ylab('') +
  theme(legend.position = "right",plot.title = element_text(hjust = 0.5,size=16), 
        axis.line = element_blank(), axis.text=element_blank(), axis.ticks = element_blank())


g1+g2+g3+g4+plot_layout(ncol=2,nrow=2,widths=c(8,8),heights=c(38,38))+plot_annotation(title = "Energy Produced per sq miles (by state)",theme = theme(plot.title = element_text(size = 36,face='bold',hjust=.5)))

```

However, it should be kept in mind that the actual energy production is represented by the first plot where California and Texas are dominating producers of energy and the second plot only slows the energy produced per sq miles for every state. 

## Correlation of Energy Consumption, CO2 Emission and Economic Indicators

The WB data validates that the CO2 emission is correlated with the fossil fuel based energy consumption in the United States. As more of our growing energy demands are met from fossil fuel based energy sources such as oil and coal – it leads to the rise of CO2 emission per capita. For example, we observed a more than 30% increase in per capita C02 emission between 1960 -1980 with about 50% increase in fossil-fuel based energy consumption.  


```{r, echo=F, warning=F}
library(tidyverse)
library (dplyr)
library(ggpmisc)

us_wb_df <- as.data.frame(read_csv('data/raw/clean_worldbank.csv'))
row.names(us_wb_df)<-us_wb_df[,"Series"]
us_wb_df <-us_wb_df[,2:length(names(us_wb_df))]

rownames(us_wb_df) <- c('Pop T', 'Pop G', 'Area', 'Pov Rt', 'GNI T','GNI PC', 'GNI PPP', 'GNI PPP PC', 'Inc Shr', 'Life Exp', 'Fertility Rt', 'Ado Ft', 'Contr Pr', 'Birth Rt', 'Mort Rt', 'Under Wt', 'Imm Mea', 'Prim Comp', 'Sec Sch', 'Prim Sch', 'HIV', 'Forest A', 'Water P', 'Energy Use PC', 'C02 Emis PC', 'Elec Cons PC', 'GDP T', 'GDP %', 'Infl %', 'Agri %', 'Industry %', 'Exports %', 'Imports %', 'Gross %', 'Revenue, %', 'Start up', 'Mkt Cap %', 'Military %', 'Mobile Sub', 'HT Exp %', 'Merchandise %', 'Net Bart', 'Ext Debt', 'Debt Sv %', 'Migration', 'Remittance', 'FDI', 'ODA')

my_us_wb_df_p <- as.data.frame(t(us_wb_df))

my_us_wb_df_p <- my_us_wb_df_p %>% relocate('GDP %', .before = 'GDP T')
my_us_wb_df_p <- my_us_wb_df_p %>% relocate('GNI PC', .before = 'GDP %')

# Annual CO2 emission trend plot

ggplot(my_us_wb_df_p, aes( my_us_wb_df_p$`Energy Use`, my_us_wb_df_p$`C02 Em`)) +
  geom_point() +
  ylab("CO2 emissions (metric tons per capita)") +
  xlab("Energy use per capita (kg of oil)") +
  ggtitle("U.S. CO2 emission correlated to fossil-fuel based \n energy consumption 1960 - 2020") +
  geom_smooth( method = 'lm', se = FALSE) +
  stat_poly_eq(formula = y~x) +
  theme_classic(16) +
  theme(plot.title = element_text(hjust=.5))
```

The parallel coordinate plot below shows the correlation of per capita CO2 emission with the per capita fossil fuel based power consumption, overall electricity consumption and gross national income (GNI) along with gross domestic product (GDP %) change. The crossing lines between CO2 emission and electricity consumption represent the negative correlation of these factors as more electricity production shifted to clean and renewable energy sources. The electricity consumption maintained a positive correlation with GNI.    

```{r, echo=F, warning=F}

library(GGally)


ggparcoord(my_us_wb_df_p, columns = 23:27,  alphaLines = 0.3, scale = "uniminmax") + 
  geom_vline(xintercept = 1:8, color = "lightblue") +
  ggtitle("U.S. Macro Energy Consumption, \n CO2 Emission, & Economic Indicators ") + 
  xlab("") + 
  ylab("Standardize Scale") +
  theme_classic(16) +
  theme(plot.title = element_text(hjust=.5))



```

We observed the clear reduction in per capita CO2 emission of about 25% between 2000 - 2020 as more and more energy demands were met from clean and renewable energy sources, such as wind and solar, as well as natural gas, a significantly cleaner fossil-fuel compared to coal.

```{r, echo=F, warning=F}

library(tidyverse)
library (dplyr)
library(ggpmisc)


my_us_wb_df_p_yr <- my_us_wb_df_p
my_us_wb_df_p_yr <-tibble::rownames_to_column(my_us_wb_df_p_yr, "Year")
my_us_wb_df_p_yr$Year <- as.numeric(my_us_wb_df_p_yr$Year)
my_us_wb_df_p_yr_2000 <-my_us_wb_df_p_yr %>% filter(Year > 1999)

ggplot(my_us_wb_df_p_yr_2000, aes( my_us_wb_df_p_yr_2000$Year, my_us_wb_df_p_yr_2000$`C02 Em`)) +
  geom_point() +
  ylab("CO2 emissions (metric tons per capita)") +
  xlab("Year") +
  ggtitle("U.S. reversed the CO2 per capita \n emission trend in 2000 - 2020") +
  stat_smooth(colour = "green") +
  theme_classic(16) +
  theme(plot.title = element_text(hjust=.5))

```




```{r}

# Correlation between Energy ETF and macro trends in energy

min_max_norm <- function(x) {
    (x - min(x)) / (max(x) - min(x))
  }

oil <- read_csv('data/clean/clean_crushing_crude_oil.csv')
oil$Day <- as.Date(oil$Day,format = '%m/%d/%Y')
names(oil) <- c("Date","Cost")


#cnrg <-read_csv('data/raw/CNRG.csv')
#icln <- read_csv('data/raw/ICLN.csv')
xle <- read_csv('data/raw/XLE.csv')



#cnrg$Close<-as.data.frame(lapply(cnrg[5],min_max_norm))$Close
#icln$Close<-as.data.frame(lapply(icln[5],min_max_norm))$Close
xle$Close<-as.data.frame(lapply(xle[5],min_max_norm))$Close


energy_df <- df  %>% 
  group_by(fuel, date) %>%
  summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE)) %>%
  filter(fuel %in% c('all_solar', 'coal','natural_gas', 'wind', 'nuclear', 'conventional_hydroelectric'))

total_energy <- energy_df %>% 
  group_by(date) %>% 
  summarise(thousand_megawatthours = sum(thousand_megawatthours, na.rm = TRUE))

total_energy$thousand_megawatthours<-as.data.frame(lapply(total_energy[2],min_max_norm))$thousand_megawatthours

names(total_energy) <- c('Date','thousand_megawatthours')

colors<-c("Total enegry (normalized)" = "red", "Normalized Closing Price (XLE)" = "blue")
merged_data <-merge(xle,total_energy,by='Date')

g1<-ggplot(merged_data)+
  geom_line(aes(Date,thousand_megawatthours,color="Total enegry (normalized)"))+
  geom_line(aes(Date,Close,color="Normalized Closing Price (XLE)"))+
  labs(x='Year', y='normalized scale',color='Legend')


df_solar2 <- df_all_solar %>% 
  group_by(date) %>% 
  summarize(solar_megawatt = sum(thousand_megawatthours)) 

df_solar2$solar_megawatt<-as.data.frame(lapply(df_solar2[2],min_max_norm))$solar_megawatt

df_coal2 <- df_coal %>% 
  group_by(date) %>% 
  summarize(coal_megawatt = sum(thousand_megawatthours)) 

df_coal2$coal_megawatt<-as.data.frame(lapply(df_coal2[2],min_max_norm))$coal_megawatt

df_natural_gas2 <- df_natural_gas %>% 
  group_by(date) %>% 
  summarize(gas_megawatt = sum(thousand_megawatthours)) 

df_natural_gas2$gas_megawatt<-as.data.frame(lapply(df_natural_gas2[2],min_max_norm))$gas_megawatt

df_wind2 <- df_wind %>% 
  group_by(date) %>% 
  summarize(wind_megawatt = sum(thousand_megawatthours)) 

df_wind2$wind_megawatt<-as.data.frame(lapply(df_wind2[2],min_max_norm))$wind_megawatt

names(xle)<-c("date","Open","High","Low","Close","Adj Close","Volume")

colors<-c("solar"="goldenrod3","coal"="grey1","natural gas"="darkolivegreen3","wind"="dodgerblue3","close price (XLE)"="red")

merged_data1 <-merge(xle,df_solar2,by='date')
merged_data2 <-merge(merged_data1,df_coal2,by='date')
merged_data3 <-merge(merged_data2,df_natural_gas2,by='date')
merged_data <-merge(merged_data3,df_wind2,by='date')

g2<-ggplot(merged_data)+
  geom_line(aes(date,solar_megawatt,color="solar"))+
  geom_line(aes(date,coal_megawatt,color="coal"))+
  geom_line(aes(date,gas_megawatt,color="natural gas"))+
  geom_line(aes(date,wind_megawatt,color="wind"))+
  geom_line(aes(date,Close,color="close price (XLE)"))+
  labs(x='Year', y='normalized scale',color='Legend')+
  scale_color_manual(values=colors)


# g1+g2+plot_layout(nrow=2,ncol=1)

```


```{r}
df1 <- df  %>% 
  group_by(fuel, date) %>% 
  summarise(mil_mwhrs = sum(thousand_megawatthours, na.rm = TRUE)/1000) %>% 
  filter(fuel %in% c('all_solar', 'coal','natural_gas', 'wind', 'nuclear', 'conventional_hydroelectric')) %>%
  mutate(fuel = replace(fuel, fuel == "conventional_hydroelectric", "hydroelectric")) %>% 
  mutate(fuel = replace(fuel, fuel == "all_solar", "solar"))


df2<-merge(df1,xle,by='date')

  # ggplot(df2) +
  # geom_line(aes(x = date, y = mil_mwhrs,color = fct_relevel(fuel, c('natural_gas', 'coal','nuclear', 'hydroelectric', 'wind', 'solar')))) +
  # geom_line(aes(x=date,y=Close*175))+
  # ggtitle('USA Net Generation by Fuel Type') + 
  # xlab("Date") +
  # ylab("MW Hours (Mil)") +
  # theme_classic(12)+
  # theme(legend.title=element_blank(), plot.title = element_text(hjust=.5))
```

```{r}
all_energy <- df  %>% 
  group_by(fuel, date) %>% 
  summarise(mil_mwhrs = sum(thousand_megawatthours, na.rm = TRUE)/1000) %>% 
  filter(fuel %in% c('all_solar', 'coal','natural_gas', 'wind', 'nuclear', 'conventional_hydroelectric')) %>%
  mutate(fuel = replace(fuel, fuel == "conventional_hydroelectric", "hydroelectric")) %>% 
  mutate(fuel = replace(fuel, fuel == "all_solar", "solar")) %>%
  pivot_wider(names_from=fuel, id_cols = c(date), values_from = mil_mwhrs)

all_data <-merge(all_energy,xle,by='date')

library(ggpmisc)

# ggplot(all_data,aes(x=solar,y=Close))+
#   geom_point()+
#   geom_smooth(method = "lm", se = TRUE)+
#   stat_poly_eq(formula=y~x)
# 
# ggplot(all_data,aes(x=coal,y=Close))+
#   geom_point()+
#   geom_smooth(method = "lm", se = TRUE)+
#   stat_poly_eq(formula=y~x)
# 
# ggplot(all_data,aes(x=natural_gas,y=Close))+
#   geom_point()+
#   geom_smooth(method = "lm", se = TRUE)+
#   stat_poly_eq(formula=y~x)
# 
# ggplot(all_data,aes(x=solar,y=Close))+
#   geom_point()+
#   geom_smooth(method = "lm", se = TRUE)+
#   stat_poly_eq(formula=y~x)
```
