==> 06-interactive.Rmd <==
# Interactive component

The bar chart given below illustrates the % of energy that is renewable in every state for a given year. The interactive component of the the plot is the date slider which allows the user to move change the year for which the plot has been created. The plot allows the user to see which states are more dependent on renewable energy and see how it changes over time. One interesting observation is Idaho (ID) is always among the top three states. This is because the major power source for Idaho is hydroelectirc power which is a form of renewable energy. The plot was created by referencing the follow plots ([link](https://bl.ocks.org/martinjc/f2241a09bd18caad10fc7249ca5d7816) to plot 1 and [link](https://bl.ocks.org/martinjc/8a2a545d5c48ef1ff65b747554332e3d) to plot 2).

<iframe src="draft code/interactive.html" width="700" height="810"></iframe>
